Support new config format and upgrade system (#89)
This PR is meant to update the starter kit algos to handle new season 5 changes involving a much more robust config format, as well as a new upgrade system that lets you upgrade your towers a single time to a stronger version.

Note, for local development you can change these variables in the new config to allow your old algos to play with your new algos: "seasonCompatibilityModeP1" "seasonCompatibilityModeP2". So if your P1 algo is old, set seasonCompatibilityModeP1 to 4, and if your P2 algo is new set seasonCompatibilityModeP2 to 5. These numbers correspond to the season. You can upload this modified config to the playground to play with your old algos there as well. Note new algos are not backwards compatible need to be played with seasonCompatibilityMode as 5.

## Python Algo Additional Notes:
 - CORES and BITS variables are switched to match everything else (cores is 0, bits is 1).
 - Costs are now returned as a list[] of two numbers, one for CORES one for BITS to support configs that have units that use both resources.
 - GameUnit class is also modified to have different fields.

## Java algo notes:
 - A ton of behind the scenes changes. 
 - Removal of various enums and classes separating firewalls and information units. Unified Unit class that contains UnitInformation.
 - Algo's strategy now upgraded and matches python algo. A few new helper functions added as well.

## Rust algo notes:
 - Serialization behind the scenes completely revamped.
 - Doesn't have as many helper functions as other algos yet. Need to calculate changes to unit stats based on upgrades yourself for example.

P.S. The entire engine code was refactored to facilitate these features and future development, it was rigorously tested but let us know if you notice any strange behavior from the new game engine.